% 
% Copyright (c) 2010, 2015, IBM Corp. All rights reserved. 
% 		
% This program is free software: you can redistribute it and/or modify 
% it under the terms of the GNU General Public License as published by 
% the Free Software Foundation, either version 3 of the License, or 
% (at your option) any later version. 
%
% This program is distributed in the hope that it will be useful, 
% but WITHOUT ANY WARRANTY; without even the implied warranty of 
% MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the 
% GNU General Public License for more details. 
%
% You should have received a copy of the GNU General Public License 
% along with this program. If not, see <http://www.gnu.org/licenses/>. 
% 
% 
\name{ida.col.def methods}
\docType{methods}
\alias{!,ida.col.def-method}
\alias{as.character,ida.col.def-method}
\alias{as.integer,ida.col.def-method}
\alias{as.numeric,ida.col.def-method}
\alias{as.vector,ida.col.def-method}
\alias{as.vector,ida.col.def,ANY-method}
\alias{ifelse,ida.col.def-method}
\alias{print,ida.col.def-method}
\alias{db.is.null}
\alias{format.ida.col.def}

\title{Available methods for ida.col.def}
\description{
\code{ida.col.def} objects are used to define new columns of a \code{ida.data.frame} based
on existing ones.

For details see the documentation of \code{\link{ida.data.frame}}.

}